# synapse/src/main/java/com/synapse/synapse/auth/AuthenticationController.java
package com.synapse.synapse.auth;

import com.synapse.synapse.auth.request.ActivationRequest;
import com.synapse.synapse.auth.request.AuthenticationRequest;
import com.synapse.synapse.auth.request.RefreshRequest;
import com.synapse.synapse.auth.request.RegistrationRequest;
import com.synapse.synapse.auth.response.AuthenticationResponse;
import io.swagger.v3.oas.annotations.tags.Tag;
import jakarta.validation.Valid;
import lombok.RequiredArgsConstructor;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.PostMapping;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RestController;

@RestController
@RequestMapping("/api/v1/auth")
@RequiredArgsConstructor
@Tag(name = "Authentication", description = "Authentication API")
public class AuthenticationController {

    private final AuthenticationService service;

    @PostMapping("/login")
    public ResponseEntity<AuthenticationResponse> login(
            @Valid
            @RequestBody
            final AuthenticationRequest request) {
        return ResponseEntity.ok(this.service.login(request));
    }

    @PostMapping("/register")
    public ResponseEntity<Void> register(
            @Valid
            @RequestBody
            final RegistrationRequest request) {
        this.service.register(request);
        return ResponseEntity.status(HttpStatus.CREATED)
                .build();
    }

    @PostMapping("/refresh")
    public ResponseEntity<AuthenticationResponse> refresh(
            @RequestBody
            final RefreshRequest req) {
        return ResponseEntity.ok(this.service.refreshToken(req));
    }

    @PostMapping("/activate")
    public ResponseEntity<Void> activateAccount(
            @Valid @RequestBody ActivationRequest request
    ) {
        service.activateAccount(request);
        return ResponseEntity.ok().build();
    }
}